requirements.txt
pyaudio
synthesizer
openai
python-dotenv
